worker_cluster/requirements.txt
redis==5.0.4
python-dotenv==1.0.1
psutil==5.9.8
httpx==0.27.0
